Refactor bridges and tests
