+++ etc/travis/suites/packages/script.sh
@@ -14,10 +14,8 @@ run_phpstan() {
       return 0
   fi
 
-  PHPSTAN_LEVEL=1
-
   print_header "Testing (PHPStan) ${package}" "Monofony"
-  run_command "(cd ${package_dir} && vendor/bin/phpstan analyse -c phpstan.neon -l ${PHPSTAN_LEVEL} ./)"
+  run_command "make test-package-phpstan path=src/Monofony/${package}"
 }
 
 clean_package() {
