+++ src/Monofony/Bridge/Behat/phpstan.neon
@@ -1,3 +1,5 @@
 parameters:
+    paths:
+        - ./
     excludes_analyse:
         - 'vendor/**.php'
