+++ src/Monofony/Bridge/FOSOAuthServer/LICENSE
