+++ src/Monofony/Bridge/SyliusUser/phpstan.neon
@@ -1,4 +1,6 @@
 parameters:
+    paths:
+        - ./
     excludes_analyse:
         - 'vendor/**.php'
-        - '**/spec/**.php'
+        - 'spec/**.php'
