+++ src/Monofony/Bundle/CoreBundle/Resources/config/services.yaml
@@ -10,5 +10,5 @@ services:
         tags:
             - { name: data_collector, template: '@MonofonyCore/DataCollector/monofony.html.twig', id: monofony_core, priority: -513 }
 
-    Monofony\Bridge\FOS\OAuthServer\Entity\ClientManager:
+    Monofony\Bridge\FOSOAuthServer\Entity\ClientManager:
         $class: '%fos_oauth_server.model.client.class%'
